fix: cast to int

+++ src/config/settings.py
@@ -125,7 +125,7 @@ REST_FRAMEWORK = {
 MINUTE = 60
 HOUR = MINUTE * 60
 
-VIEW_CACHE_TTL = MINUTE if DEBUG else os.environ.get("VIEW_CACHE_TTL", 3 * HOUR)
+VIEW_CACHE_TTL = MINUTE if DEBUG else int(os.environ.get("VIEW_CACHE_TTL", 3 * HOUR))
 
 # spire.scraper
